mob experiments

+++ src/main/java/com/teamcqr/chocolatequestrepoured/objects/entity/EBaseHealths.java
@@ -2,29 +2,29 @@ package com.teamcqr.chocolatequestrepoured.objects.entity;
 
 public enum EBaseHealths {
 	
-	NPC(20.0D),
-	WALKER(40.0D),
-	PIGMEN(50.0D),
-	MACHINES(35.0D),
-	UNDEAD(25.0D),
-	MINOTAURS(30.0D),
-	DWARVES(30.0D),
-	ORCS(30.0D),
-	GREMLINS(15.0D),
-	OGRES(35.0D),
-	BEES(15.0D),
-	MOUNTS(25.0D),
-	SPECTRES(30.0D),
-	PIRATES(25.0D),
-	MANDRILS(30.0D);
+	NPC(20.0F),
+	WALKER(40.0F),
+	PIGMEN(50.0F),
+	MACHINES(35.0F),
+	UNDEAD(25.0F),
+	MINOTAURS(30.0F),
+	DWARVES(30.0F),
+	ORCS(30.0F),
+	GREMLINS(15.0F),
+	OGRES(35.0F),
+	BEES(15.0F),
+	MOUNTS(25.0F),
+	SPECTRES(30.0F),
+	PIRATES(25.0F),
+	MANDRILS(30.0F);
 	
-	private double baseHealth;
+	private float baseHealth;
 	
-	private EBaseHealths(double baseHealth) {
+	private EBaseHealths(float baseHealth) {
 		this.baseHealth = baseHealth;
 	}
 	
-	public double getValue() {
+	public float getValue() {
 		return baseHealth;
 	}
